requirements.txt
Django>=5.2.9
djangorestframework>=3.14.0
drf-yasg>=1.21.7
Pillow>=10.0.0
django-filter>=23.2
